requirement.txt
Streamlit
Pillow
Numpy
Plotly
Ultralytics
